8.14
8.14까지 강의 시청 후 구현까지 완료함

+++ src/controllers/videoController.js
@@ -24,21 +24,38 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const id = req.params.id;
+  const {
+    session: {
+      user: { _id },
+    },
+  } = req;
+
   const video = await Video.findById(id);
   if (!video) {
-    // 에러가 발생하는 경우를 처리하는 if문
     return res.status(404).render("404", { pageTitle: "Video Not Found." });
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   res.render("Edit", { pageTitle: `Edit ${video.title}`, video: video });
 };
 
 export const postEdit = async (req, res) => {
+  const {
+    session: {
+      user: { _id },
+    },
+  } = req;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     // 에러가 발생하는 경우를 처리하는 if문
     return res.status(404).render("404", { pageTitle: "Video Not Found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -80,7 +97,19 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+  const {
+    session: {
+      user: { _id },
+    },
+  } = req;
   const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video Not Found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
